+++ plugins/polygonMask/src/index.ts
@@ -55,7 +55,7 @@ export async function loadPolygonMaskPlugin(tsParticles: Main): Promise<void> {
 
     const plugin = new Plugin();
 
-    tsParticles.addPlugin(plugin);
+    await tsParticles.addPlugin(plugin);
 }
 
 export * from "./Enums";