+++ shapes/star/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { StarDrawer } from "./StarDrawer";
 
-export function loadStarShape(tsParticles: Main): void {
-    tsParticles.addShape("star", new StarDrawer());
+export async function loadStarShape(tsParticles: Main): Promise<void> {
+    await tsParticles.addShape("star", new StarDrawer());
 }