+++ shapes/text/src/index.ts
@@ -1,10 +1,10 @@
 import type { Main } from "tsparticles-engine";
 import { validTypes, TextDrawer } from "./TextDrawer";
 
-export function loadTextShape(tsParticles: Main): void {
+export async function loadTextShape(tsParticles: Main): Promise<void> {
     const drawer = new TextDrawer();
 
     for (const type of validTypes) {
-        tsParticles.addShape(type, drawer);
+        await tsParticles.addShape(type, drawer);
     }
 }