+++ updaters/size/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { SizeUpdater } from "./SizeUpdater";
 
-export function loadSizeUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("size", () => new SizeUpdater());
+export async function loadSizeUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("size", () => new SizeUpdater());
 }