+++ updaters/wobble/src/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "tsparticles-engine";
 import { WobbleUpdater } from "./WobbleUpdater";
 
-export function loadWobbleUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("wobble", (container) => new WobbleUpdater(container));
+export async function loadWobbleUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("wobble", (container) => new WobbleUpdater(container));
 }